+++ CHANGELOG.md
@@ -40,7 +40,7 @@ request adding CHANGELOG notes for breaking (!) changes and possibly other secti
 
 ### Breaking changes
 
-- The default request ID header name has changed from `Polaris-Request-Id` to `x-request-id`.
+- The default request ID header name has changed from `Polaris-Request-Id` to `X-Request-ID`.
 
 ### New Features